// pacific-sdk/src/code/types.ts
export type EmanationSDKScreen = {
  meta: {
    theme: EmanationSDKTheme;
  };
  name: string; // E.g. "Profile"
  description: string; // E.g. "The profile screen"
  root: WorldviewNode;
};

export type EmanationSDKApp = {
  meta: {};

  name: string; // E.g. "Finance app"
  description: string; // E.g. "A personal finance app"
  screens: EmanationSDKScreen[];
};

export type EmanationSDKEpoch = {
  name: string; // E.g. "" + Date.now()
  apps: EmanationSDKApp[];
};

export type EmanationSDKAssets = {
  [key: string]: number[] | string;
};